Move components to dedicated file
