+++ app/javascript/src/apps/mydb/elements/details/reactions/variationsTab/ReactionVariations.js
@@ -1,7 +1,7 @@
 /* eslint-disable react/display-name */
 import { AgGridReact } from 'ag-grid-react';
 import React, {
-  useRef, useState, useEffect, useCallback
+  useRef, useState, useCallback
 } from 'react';
 import {
   Button, OverlayTrigger, Tooltip, Alert,
@@ -12,8 +12,7 @@ import PropTypes from 'prop-types';
 import Reaction from 'src/models/Reaction';
 import {
   createVariationsRow, copyVariationsRow, updateVariationsRow, getCellDataType,
-  temperatureUnits, durationUnits, getStandardUnit, materialTypes, updateColumnDefinitions,
-  getUserFacingUnit
+  temperatureUnits, durationUnits, getStandardUnit, materialTypes
 } from 'src/apps/mydb/elements/details/reactions/variationsTab/ReactionVariationsUtils';
 import {
   AnalysesCellRenderer, AnalysesCellEditor, getReactionAnalyses, updateAnalyses, getAnalysesOverlay, AnalysisOverlay
@@ -27,160 +26,9 @@ import {
   PropertyFormatter, PropertyParser,
   MaterialFormatter, MaterialParser,
   EquivalentFormatter, EquivalentParser,
-  RowToolsCellRenderer, NoteCellRenderer, NoteCellEditor
+  NoteCellRenderer, NoteCellEditor,
+  RowToolsCellRenderer, MenuHeader
 } from 'src/apps/mydb/elements/details/reactions/variationsTab/ReactionVariationsCellComponents';
-
-function MenuHeader({
-  column, context, setSort, names
-}) {
-  const { field, entryDefs } = column.colDef;
-  const { columnDefinitions, setColumnDefinitions } = context;
-  const [ascendingSort, setAscendingSort] = useState('inactive');
-  const [descendingSort, setDescendingSort] = useState('inactive');
-  const [noSort, setNoSort] = useState('inactive');
-  const [name, setName] = useState(names[0]);
-  const { currentEntry, displayUnit, availableEntriesWithUnits } = entryDefs;
-  const [entry, setEntry] = useState(currentEntry);
-  const [unit, setUnit] = useState(displayUnit);
-  const [units, setUnits] = useState(availableEntriesWithUnits[currentEntry]);
-
-  const onSortChanged = () => {
-    setAscendingSort(column.isSortAscending() ? 'sort_active' : 'inactive');
-    setDescendingSort(column.isSortDescending() ? 'sort_active' : 'inactive');
-    setNoSort(
-      !column.isSortAscending() && !column.isSortDescending()
-        ? 'sort_active'
-        : 'inactive'
-    );
-  };
-
-  useEffect(() => {
-    column.addEventListener('sortChanged', onSortChanged);
-    onSortChanged();
-  }, []);
-
-  const onSortRequested = (order, event) => {
-    setSort(order, event.shiftKey);
-  };
-
-  const onUnitChanged = () => {
-    const newUnit = units[(units.indexOf(unit) + 1) % units.length];
-    const newColumnDefinitions = updateColumnDefinitions(
-      columnDefinitions,
-      field,
-      'entryDefs',
-      { currentEntry: entry, displayUnit: newUnit, availableEntriesWithUnits }
-    );
-
-    setUnit(newUnit);
-    setColumnDefinitions(newColumnDefinitions);
-  };
-
-  const unitSelection = (
-    <Button
-      className={`unitSelection ${entry === 'equivalent' ? 'd-none' : 'd-inline'}`}
-      variant="success"
-      size="sm"
-      onClick={onUnitChanged}
-    >
-      {getUserFacingUnit(unit)}
-    </Button>
-  );
-
-  const onEntryChanged = () => {
-    const entryKeys = Object.keys(availableEntriesWithUnits);
-    const newEntry = entryKeys[(entryKeys.indexOf(entry) + 1) % entryKeys.length];
-    const newUnits = availableEntriesWithUnits[newEntry];
-    const newUnit = newUnits[0];
-    let newColumnDefinitions = updateColumnDefinitions(
-      columnDefinitions,
-      field,
-      'cellDataType',
-      getCellDataType(newEntry)
-    );
-    newColumnDefinitions = updateColumnDefinitions(
-      newColumnDefinitions,
-      field,
-      'entryDefs',
-      { currentEntry: newEntry, displayUnit: newUnit, availableEntriesWithUnits }
-    );
-
-    setEntry(newEntry);
-    setUnits(newUnits);
-    setUnit(newUnit);
-    setColumnDefinitions(newColumnDefinitions);
-  };
-
-  const entrySelection = (
-    <Button
-      className={`entrySelection ${['temperature', 'duration'].includes(entry) ? 'd-none' : 'd-inline'}`}
-      variant="light"
-      size="sm"
-      disabled={Object.keys(availableEntriesWithUnits).length === 1}
-      onClick={onEntryChanged}
-    >
-      {entry}
-    </Button>
-  );
-
-  const sortMenu = (
-    <div className="sortHeader d-flex align-items-center">
-      <div
-        onClick={(event) => onSortRequested('asc', event)}
-        onTouchEnd={(event) => onSortRequested('asc', event)}
-        className={`customSortDownLabel ${ascendingSort}`}
-      >
-        <i className="fa fa-chevron-up fa-fw" />
-      </div>
-      <div
-        onClick={(event) => onSortRequested('desc', event)}
-        onTouchEnd={(event) => onSortRequested('desc', event)}
-        className={`customSortUpLabel ${descendingSort}`}
-      >
-        <i className="fa fa-chevron-down fa-fw" />
-      </div>
-      <div
-        onClick={(event) => onSortRequested('', event)}
-        onTouchEnd={(event) => onSortRequested('', event)}
-        className={`customSortRemoveLabel ${noSort}`}
-      >
-        <i className="fa fa-times fa-fw" />
-      </div>
-    </div>
-  );
-
-  return (
-    <div className="d-grid">
-      <span
-        className="header-title"
-        onClick={() => setName(names[(names.indexOf(name) + 1) % names.length])}
-      >
-        {name}
-      </span>
-      <div>
-        {entrySelection}
-        {' '}
-        {unitSelection}
-      </div>
-      {sortMenu}
-    </div>
-  );
-}
-
-MenuHeader.propTypes = {
-  column: PropTypes.shape({
-    colDef: PropTypes.object.isRequired,
-    isSortAscending: PropTypes.func.isRequired,
-    isSortDescending: PropTypes.func.isRequired,
-    addEventListener: PropTypes.func.isRequired,
-  }).isRequired,
-  context: PropTypes.shape({
-    columnDefinitions: PropTypes.arrayOf(PropTypes.object).isRequired,
-    setColumnDefinitions: PropTypes.func.isRequired,
-  }).isRequired,
-  setSort: PropTypes.func.isRequired,
-  names: PropTypes.arrayOf(PropTypes.string).isRequired,
-};
 
 export default function ReactionVariations({ reaction, onReactionChange }) {
   const gridRef = useRef(null);
